Folds, includes and shennanigans

+++ MiniSumo_2012/delay_ms.h
@@ -2,10 +2,11 @@
  * File Name : delay_ms.h
  * Purpose : Longer delays
  * Creation Date : 26-08-2012
- * Last Modified : sön 26 aug 2012 16:04:24
+ * Last Modified : tor 30 aug 2012 13:54:35
  * Created By : Gabriel Fornaeus, <gf@hax0r.se>
  *
  */
 
 #include <util/delay.h>
-void delay_ms(uint16_t ms);
+#include <stdint.h>
+void delay_ms(uint8_t ms);
